+++ src/posts/posts.service.ts
@@ -1,11 +1,22 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Inject,
+  Injectable,
+  NotFoundException,
+  forwardRef,
+} from '@nestjs/common';
 import { CreatePosts } from './dtos/CreatePosts';
 import { UpdatePosts } from './dtos/UpdatePosts';
 import { PostsRepository } from './posts.repository';
+import { PublicationsService } from 'src/publications/publications.service';
 
 @Injectable()
 export class PostsService {
-  constructor(private readonly postRepository: PostsRepository) {}
+  constructor(
+    private readonly postRepository: PostsRepository,
+    @Inject(forwardRef(() => PublicationsService))
+    private readonly publicationsService: PublicationsService,
+  ) {}
 
   async post(CreatePost: CreatePosts) {
     return await this.postRepository.post(CreatePost);
@@ -26,6 +37,9 @@ export class PostsService {
   }
 
   async delete(id: number) {
+    await this.getById(id);
+    const publication = await this.publicationsService.getByPostId(id);
+    if (publication) throw new ForbiddenException();
     return await this.postRepository.delete(id);
   }
 }
